+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 #### Next Release
 - Updated Moshi to 1.3.0.
+- Highlighted query in search results.
 
 #### v1.14.6 (2016-10-18)
 - Downgraded Moshi to 1.2.0.